细节更新

+++ controller/team.go
@@ -1,23 +1,30 @@
 package controller
 
 import (
-	"github.com/gin-gonic/gin"
+	"fmt"
+	"math/rand"
+	"strconv"
+	"time"
 	"walk-server/model"
 	"walk-server/utility"
 	"walk-server/utility/initial"
+
+	"github.com/gin-gonic/gin"
 )
 
 // CreateTeamData 接收创建团队信息的数据类型
 type CreateTeamData struct {
-	Name     string `json:"name" binding:"required"`
-	Password string `json:"password" binding:"required"`
-	Route    uint8  `json:"route" binding:"required"`
+	Name       string `json:"name" binding:"required"`
+	Route      uint8  `json:"route" binding:"required"`
+	Password   string `json:"password" binding:"required"`
+	AllowMatch bool   `json:"allow_match"`
 }
 
 // UpdateTeamData 更新团队信息的数据类型
 type UpdateTeamData struct {
-	Name  string `json:"name"`
-	Route uint8  `json:"route"`
+	Name       string `json:"name" binding:"required"`
+	Route      uint8  `json:"route" binding:"required"`
+	AllowMatch bool   `json:"allow_match"`
 }
 
 // JoinTeamData 加入团队时接收的信息类型
@@ -27,7 +34,6 @@ type JoinTeamData struct {
 }
 
 func CreateTeam(context *gin.Context) {
-	// TODO 加入对当天团队数上限的判断
 	// 获取 jwt 数据
 	jwtToken := context.GetHeader("Authorization")[7:]
 	jwtData, _ := utility.ParseToken(jwtToken)
@@ -47,6 +53,7 @@ func CreateTeam(context *gin.Context) {
 
 	if person.Status != 0 { // 现在已经加入了一个团队
 		utility.ResponseError(context, "请先退出或解散原来的团队")
+		return
 	}
 
 	if person.CreatedOp == 0 {
@@ -54,20 +61,22 @@ func CreateTeam(context *gin.Context) {
 	} else {
 		// 再数据库中插入一个团队
 		team := model.Team{
-			Name:      createTeamData.Name,
-			Num:       1,
-			Password:  createTeamData.Password,
-			Captain:   person.Name,
-			Route:     createTeamData.Route,
-			Submitted: false,
+			Name:       createTeamData.Name,
+			Num:        1,
+			AllowMatch: createTeamData.AllowMatch,
+			Password:   createTeamData.Password,
+			Captain:    person.OpenId,
+			Route:      createTeamData.Route,
+			Submitted:  false,
 		}
 		initial.DB.Create(&team)
 
 		// 将入团队后对应的状态更新
 		person.CreatedOp -= 1
-		person.Status = 1
+		person.Status = 2
+		person.TeamId = int(team.ID)
 
-		initial.DB.Model(&person).Updates(person) // 将新的用户信息写入数据库
+		initial.DB.Save(&person) // 将新的用户信息写入数据库
 
 		// 返回新的 team_id 和 jwt 数据
 		utility.ResponseSuccess(context, gin.H{
@@ -109,7 +118,7 @@ func JoinTeam(context *gin.Context) {
 		utility.ResponseError(context, "找不到团队")
 		return
 	}
-	if team.Submitted == true {
+	if team.Submitted {
 		utility.ResponseError(context, "该队伍已提交，无法加入")
 		return
 	}
@@ -125,6 +134,7 @@ func JoinTeam(context *gin.Context) {
 	} else {
 		person.Status = 1
 		person.JoinOp--
+		person.TeamId = int(team.ID)
 		initial.DB.Model(&person).Updates(person) // 将新的用户信息写入数据库
 		utility.ResponseSuccess(context, nil)
 	}
@@ -153,14 +163,16 @@ func GetTeamInfo(context *gin.Context) {
 	var persons []model.Person
 	var leader model.Person
 	var members []gin.H
-	initial.DB.Where("team_id = ?", person.TeamId).First(&persons)
+	initial.DB.Where("team_id = ?", person.TeamId).Find(&persons)
 	for _, person := range persons {
 		if person.Status == 2 { // 队长
 			leader = person
 		} else {
 			members = append(members, gin.H{
-				"name":   person.Name,
-				"gender": person.Gender,
+				"name":    person.Name,
+				"gender":  person.Gender,
+				"open_id": person.OpenId,
+				"campus":  person.Campus,
 				"contact": gin.H{
 					"qq":     person.Qq,
 					"wechat": person.Wechat,
@@ -172,12 +184,17 @@ func GetTeamInfo(context *gin.Context) {
 
 	// 返回结果
 	utility.ResponseSuccess(context, gin.H{
-		"id":    person.TeamId,
-		"name":  team.Name,
-		"route": team.Route,
+		"id":          person.TeamId,
+		"name":        team.Name,
+		"route":       team.Route,
+		"password":    team.Password,
+		"submitted":   team.Submitted,
+		"allow_match": team.AllowMatch,
 		"leader": gin.H{
-			"name":   leader.Name,
-			"gender": leader.Gender,
+			"name":    leader.Name,
+			"gender":  leader.Gender,
+			"campus":  leader.Campus,
+			"open_id": leader.OpenId,
 			"contact": gin.H{
 				"qq":     leader.Qq,
 				"wechat": leader.Wechat,
@@ -209,9 +226,14 @@ func DisbandTeam(context *gin.Context) {
 	var team model.Team
 	initial.DB.Where("id = ?", person.TeamId).First(&team)
 
+	if team.Submitted == true {
+		utility.ResponseError(context, "该队伍已提交，无法解散")
+		return
+	}
+
 	// 查找团队所有用户
 	var persons []model.Person
-	initial.DB.Where("team_id = ?", person.TeamId).First(&persons)
+	initial.DB.Where("team_id = ?", person.TeamId).Find(&persons)
 
 	// 删除团队记录
 	initial.DB.Delete(&team)
@@ -222,6 +244,8 @@ func DisbandTeam(context *gin.Context) {
 		person.TeamId = -1
 		initial.DB.Save(&person)
 	}
+
+	utility.ResponseSuccess(context, nil)
 }
 
 func LeaveTeam(context *gin.Context) {
@@ -270,9 +294,12 @@ func RemoveMember(context *gin.Context) {
 	memberRemovedOpenID := context.Query("openid")
 
 	var personRemoved model.Person
-	result := initial.DB.Where("open_id = ?", memberRemovedOpenID).First(personRemoved)
+	result := initial.DB.Where("open_id = ?", memberRemovedOpenID).First(&personRemoved)
 	if result.RowsAffected == 0 {
 		utility.ResponseError(context, "没有这个用户")
+		return
+	} else if personRemoved.TeamId != person.TeamId {
+		utility.ResponseError(context, "不能移除别的队伍的人")
 		return
 	}
 
@@ -310,8 +337,104 @@ func UpdateTeam(context *gin.Context) {
 
 	// 更新团队信息
 	var team model.Team
-	initial.DB.Where("team_id = ?", person.TeamId).First(&team)
+	initial.DB.Where("id = ?", person.TeamId).First(&team)
 	team.Name = updateTeamData.Name
 	team.Route = updateTeamData.Route
+	team.AllowMatch = updateTeamData.AllowMatch
 	initial.DB.Save(&team)
+	utility.ResponseSuccess(context, nil)
+}
+
+func SubmitTeam(context *gin.Context) {
+	// 获取 jwt 数据
+	jwtToken := context.GetHeader("Authorization")[7:]
+	jwtData, _ := utility.ParseToken(jwtToken)
+
+	// 查找用户
+	var person model.Person
+	initial.DB.Where("open_id = ?", jwtData.OpenID).First(&person)
+
+	// 判断用户权限
+	if person.Status == 0 {
+		utility.ResponseError(context, "请先加入队伍")
+		return
+	} else if person.Status == 1 {
+		utility.ResponseError(context, "没有修改的权限")
+		return
+	}
+
+	var team model.Team
+	var teamCount model.TeamCount
+
+	initial.DB.Where("id = ?", person.TeamId).First(&team)
+	if team.Submitted == true {
+		utility.ResponseError(context, "该队伍已经提交过了")
+	}
+
+	// 开始提交
+	initial.DB.Where("day_campus = ?", utility.GetCurrentDate()*10+team.Route).First(&teamCount)
+	key := fmt.Sprintf("teamUpperLimit.%v.%v", team.Route, utility.GetCurrentDate())
+	result := initial.DB.Model(&teamCount).Where("count < ?", initial.Config.GetInt(key)).Update("count", teamCount.Count+1)
+	if result.RowsAffected == 0 { // 队伍数量到达上限
+		utility.ResponseError(context, "队伍数量已经到达上限，无法提交")
+	} else { // 团队提交状态更新
+		if team.Num < 4 {
+			utility.ResponseError(context, "队伍人数太少")
+		} else {
+			team.Submitted = true
+			initial.DB.Save(&team)
+			utility.ResponseSuccess(context, nil)
+		}
+	}
+}
+
+// RandomMatch 随机组队
+func RandomMatch(context *gin.Context) {
+	route, err := strconv.Atoi(context.Query("route"))
+	if err != nil {
+		utility.ResponseError(context, "参数错误")
+		return
+	}
+
+	// 获取 jwt 数据
+	jwtToken := context.GetHeader("Authorization")[7:]
+	jwtData, _ := utility.ParseToken(jwtToken)
+
+	// 从数据库中读取用户信息
+	var person model.Person
+	initial.DB.Where("open_id = ?", jwtData.OpenID).Find(&person)
+
+	if person.Status != 0 { // 如果在一个团队中
+		utility.ResponseError(context, "请退出或解散原来的团队")
+		return
+	}
+
+	if person.JoinOp == 0 { // 加入次数用完了
+		utility.ResponseError(context, "没有加入次数了")
+		return
+	}
+
+	// 找到一个可以加入的随机队伍
+	var teams, randTeams []model.Team
+	initial.DB.Where("num < 6 AND route = ? AND allow_match = 1 AND submitted = 0", route).Find(&teams) // 从可能的队伍中挑选前 10 条
+	rand.Seed(time.Now().UnixNano())                                                                    // 设置随机种子
+	for n := len(teams); n > 0; n-- {
+		randIndex := rand.Intn(n)
+		randTeams = append(randTeams, teams[randIndex])
+		teams[n-1], teams[randIndex] = teams[randIndex], teams[n-1] // 将 teams[randIndex] 丢到末尾
+	}
+	for _, team := range randTeams {
+		result := initial.DB.Model(&team).Where("num < 6 AND allow_match = 1 AND submitted = 0").Update("num", team.Num+1) // 更新队伍人数
+		if result.RowsAffected != 0 {                                                                                      // 更新成功
+			person.Status = 1
+			person.JoinOp--
+			person.TeamId = int(team.ID)
+			initial.DB.Model(&person).Updates(person) // 将新的用户信息写入数据库
+			utility.ResponseSuccess(context, nil)
+
+			return
+		}
+	}
+
+	utility.ResponseError(context, "没有匹配上的队伍")
 }
